fetchDataFromCacheOrServer() の作成
